shared/requirements.txt
# shared/requirements.txt
# Complete shared dependencies

# Google Cloud (updated for your stack)
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-cloud-firestore==2.13.0
google-cloud-pubsub==2.18.4
google-cloud-core==2.4.3
google-auth==2.40.2

# Core utilities
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.4.0

# Logging (IMPORTANT!)
structlog==23.2.0
logging==0.4.9.6

# Error tracking
sentry-sdk==1.38.0

# JSON handling
orjson==3.9.10

# Date/time
python-dateutil==2.9.0.post0
pytz==2025.2

# Environment detection
python-decouple==3.8

# Data handling
pandas==2.1.3
numpy==1.24.4

# Database connections
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Retry logic
tenacity==8.2.3

# Metrics and monitoring
prometheus-client==0.19.0